# packages/api-contracts/src/openapi/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "trudskill API",
    "version": "1.0.0",
    "description": "Versioned API contracts for /api/v1 with envelope/meta and critical route coverage."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "authLogin",
        "tags": ["auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/AuthTokensResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitedError"
          },
          "default": {
            "$ref": "#/components/responses/DefaultError"
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "authRefresh",
        "tags": ["auth"],
        "responses": {
          "200": {
            "$ref": "#/components/responses/AuthTokensResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "default": {
            "$ref": "#/components/responses/DefaultError"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "authLogout",
        "tags": ["auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessFlagResponse"
          },
          "default": {
            "$ref": "#/components/responses/DefaultError"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "authMe",
        "tags": ["auth"],
        "responses": {
          "200": {
            "$ref": "#/components/responses/UserResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "default": {
            "$ref": "#/components/responses/DefaultError"
          }
        }
      }
    },
    "/auth/sessions": {
      "get": {
        "operationId": "authListSessions",
        "tags": ["auth"],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SessionListResponse"
          },
          "default": {
            "$ref": "#/components/responses/DefaultError"
          }
        }
      }
    },
    "/auth/sessions/{id}": {
      "delete": {
        "operationId": "authRevokeSession",
        "tags": ["auth"],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPathParam"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessFlagResponse"
          },
          "default": {
            "$ref": "#/components/responses/DefaultError"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "IdPathParam": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
    "schemas": {
      "ApiMeta": {
        "type": "object",
        "required": ["requestId", "correlationId", "timestamp"],
        "properties": {
          "requestId": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ErrorDetails": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      },
      "ApiError": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_ERROR",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "PRECONDITION_FAILED",
              "RATE_LIMITED",
              "INTERNAL_ERROR",
              "internal_error",
              "error",
              "forbidden",
              "not_found",
              "precondition_failed",
              "invalid_credentials",
              "invalid_refresh"
            ]
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDetails"
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["error", "meta"],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
      "SuccessFlag": {
        "type": "object",
        "required": ["success"],
        "properties": {
          "success": {
            "type": "boolean"
          }
        }
      },
      "AuthTokens": {
        "type": "object",
        "required": ["accessToken", "sessionId", "expiresIn"],
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "sessionId": {
            "type": "string"
          },
          "expiresIn": {
            "type": "integer"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["login", "password"],
        "properties": {
          "login": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "LogoutRequest": {
        "type": "object",
        "required": ["sessionId"],
        "properties": {
          "sessionId": {
            "type": "string"
          }
        }
      },
      "UserPublic": {
        "type": "object",
        "required": ["id", "tenantId", "login", "status", "displayName"],
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "login": {
            "type": "string"
          },
          "email": {
            "type": ["string", "null"]
          },
          "status": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          }
        }
      },
      "SessionPublic": {
        "type": "object",
        "required": ["id", "tenantId", "userId", "expiresAt"],
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "revokedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Envelope": {
        "type": "object",
        "required": ["data", "meta"],
        "properties": {
          "data": {
            "type": "object"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      }
    },
    "responses": {
      "DefaultError": {
        "description": "Default error response envelope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "UnauthorizedError": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "RateLimitedError": {
        "description": "Rate limited error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "AuthTokensResponse": {
        "description": "Auth tokens envelope",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Envelope"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AuthTokens"
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "SuccessFlagResponse": {
        "description": "Boolean success envelope",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Envelope"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/SuccessFlag"
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "UserResponse": {
        "description": "User envelope",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Envelope"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/UserPublic"
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "SessionListResponse": {
        "description": "Sessions envelope",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Envelope"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SessionPublic"
                      }
                    }
                  }
                }
              ]
            }
          }
        }
      },
...
    }
  },
...
}